Refactor DatePicker tests: remove duplicate test file and improve test structure
- Delete DatePicker.clean.spec.ts duplicate test file
- Refactor DatePicker.spec.ts with improved test organization and type safety
- Add TypeScript type annotations for DatePickerInstance
- Replace beforeEach setup with mountComponent helper function
- Simplify test assertions and remove unnecessary DOM queries
- Update test descriptions for better clarity
- Remove event listener spy tests and replace with focused unit
